StandaloneClass / LangKeys / Parse type
