+++ src/main/java/codechicken/core/gui/GuiScrollPane.java
@@ -3,7 +3,7 @@ package codechicken.core.gui;
 import java.awt.Dimension;
 import java.awt.Rectangle;
 
-import codechicken.lib.colour.Colour.LocalizedColours;
+import codechicken.lib.colour.LocalizedColours;
 import codechicken.lib.math.MathHelper;
 import codechicken.lib.vec.Rectangle4i;
