+++ src/main/resources/assets/codechickencore/lang/en_US.lang
@@ -1,1 +1,26 @@
 codechickencore.update=Version %s of %s is available
+
+codechickencore.color.buttonTextDisabled=FFA0A0A0
+codechickencore.color.buttonTextHover=FFFFFFA0
+codechickencore.color.buttonText=FFE0E0E0
+codechickencore.color.textFieldBorder=FFA0A0A0
+codechickencore.color.textFieldBackground=FF000000
+codechickencore.color.textFieldText=E0E0E0
+codechickencore.color.textFieldTextDisabled=707070
+codechickencore.color.background=FF000000
+codechickencore.color.overlayTop=FFA0A0A0
+codechickencore.color.overlayBottom=FFA0A0A0
+codechickencore.color.overlayLeft=FFA0A0A0
+codechickencore.color.overlayRight=FFA0A0A0
+codechickencore.color.scrollbarCorners=FF8B8B8B
+codechickencore.color.scrollbarTopLeft=FFF0F0F0
+codechickencore.color.scrollbarBottomRight=FF555555
+codechickencore.color.scrollbarFill=FFC6C6C6
+codechickencore.color.scrollbarGuide=FF808080
+codechickencore.color.tooltipBgStart=F0100010
+codechickencore.color.tooltipBgEnd=F0100010
+codechickencore.color.tooltipBorderStart=505000FF
+codechickencore.color.tooltipBorderEnd=5028007F
+codechickencore.color.tooltipText=FFFFFFFF
+codechickencore.color.itemQuantityText=FFFFFF
+codechickencore.color.modDescriptionText=DDDDDD
